requirements.txt
# Core ADK dependencies
google-adk>=1.19.0
google-cloud-aiplatform>=1.70.0
google-cloud-discoveryengine>=0.13.0
google-cloud-secret-manager

# Vertex AI and Google Cloud
google-cloud-logging>=3.11.0
google-cloud-firestore>=2.18.0
google-auth>=2.35.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Async support
aiohttp>=3.10.0
asyncio>=3.4.3
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.12
typing-extensions>=4.12.0

# Language processing
langdetect>=1.0.9
indic-nlp-library>=0.92

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
httpx>=0.27.0  # For FastAPI testing

# Development
black>=24.8.0
flake8>=7.1.0
mypy>=1.11.0
isort>=5.13.0

# Logging and monitoring
colorlog>=6.8.0
structlog>=24.4.0

# Optional: For local development with mock datastores
faiss-cpu>=1.8.0  # For local vector search simulation
chromadb>=0.5.0  # Alternative local vector DB

# LangChain & Semantic Search Dependencies
langchain-core>=0.3.15
langchain-google-community>=2.0.3
langchain-google-vertexai>=2.0.8